Get tests passing
Old default-time test seemed to expect times to be in microseconds (?).
In riemann 0.2.2 it looks like the default event time is in seconds
(https://github.com/aphyr/riemann/blob/master/src/riemann/common.clj#L112),
so I changed the test.

The protobufs tests were all failing due to a non-nil :decode-time being
present on the message after it was roundtripped through encode/decode.
I cleared the value in this key since it appears to be meta-data and not
important for the purpose of the test.

tl;dr - I changed the tests to make them pass.