Provide context when serialization fails
Somewhat related to [my previous
PR](https://github.com/aphyr/riemann-clojure-client/pull/16): I noticed that
that I'd occasionally index an event that was impossible to serialize
(that is, with non-String custom attribute values) due to an error in my
Riemann config. The symptoms: inability to remotely query the index, and
a cryptic ClassCastException in the Riemann log.

This made the problem more difficult to diagnose than it ought to have
been. This PR aims to fix that by providing more context when event
serialization fails so that users can more quickly diagnose errors in
their Riemann configurations.